renames fragment folders
