+++ app/src/main/java/com/walkly/walkly/ui/battles/BattlesViewModel.kt
@@ -1,13 +1,13 @@
-package com.walkly.walkly.ui.home
+package com.walkly.walkly.ui.battles
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class MapViewModel : ViewModel() {
+class BattlesViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is map Fragment"
+        value = "This is battles Fragment"
     }
     val text: LiveData<String> = _text
 }
