+++ app/src/main/java/com/walkly/walkly/ui/map/MapFragment.kt
@@ -1,5 +1,6 @@
-package com.walkly.walkly.ui.dashboard
+package com.walkly.walkly.ui.map
 
+import android.annotation.SuppressLint
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -10,22 +11,28 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import com.walkly.walkly.R
 
-class BattlesFragment : Fragment() {
 
-    private lateinit var battlesViewModel: BattlesViewModel
+@SuppressLint("Registered")
+class MapFragment : Fragment() {
+
+    private lateinit var mapViewModel: MapViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        battlesViewModel =
-            ViewModelProviders.of(this).get(BattlesViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_battkes, container, false)
-        val textView: TextView = root.findViewById(R.id.text_dashboard)
-        battlesViewModel.text.observe(this, Observer {
+        mapViewModel =
+            ViewModelProviders.of(this).get(MapViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_map, container, false)
+        val textView: TextView = root.findViewById(R.id.text_notifications)
+        mapViewModel.text.observe(this, Observer {
             textView.text = it
         })
         return root
     }
+
+
+
+
 }
