+++ app/src/main/java/com/walkly/walkly/ui/map/MapViewModel.kt
@@ -1,13 +1,13 @@
-package com.walkly.walkly.ui.dashboard
+package com.walkly.walkly.ui.map
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class BattlesViewModel : ViewModel() {
+class MapViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is battles Fragment"
+        value = "This is map Fragment"
     }
     val text: LiveData<String> = _text
 }
